+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Product price discount condition `hasBestPrice`.
 
 ## [2.7.0] - 2023-04-05
-
 ### Added
 - Seller id(s) condition `sellerId`